# Improve Performance of TTFunk::Table::OS2#group_original_code_points_by_bit