## Abstruct

fix #82

Original code takes `O(N * M)` time  where N is UNICODE_RANGES size and M is code_map keys size.
Improved code takes `O(max(N, M))` time.

## Idea

Original code calls `r.cover?` with all pair of `UNICODE_RANGES` and `code_map keys`.
This takes `O(N * M)` time.

If `UNICODE_RANGES` and `code_map keys` are both sorted, first `UNICODE_RANGE` contains some `code_map keys` of the head flagment (or empty), next `UNICODE_RANGE` contains next flagment, and so on.
This storategy loops `UNICODE_RANGES` and `code_map keys` indivisualy, so it takes `O(max(N, M))` time.

## Performance Test Result

I measure `Problem: TTFunk::Subset#encode` time with #82 reproduction code.(English Font changed to 'spec/fonts/DejaVuSans.ttf')
It's about 8x faster than master branch, still 4x slower than 1.5.1 branch.

|branch|font|time[s]|
|---|---|---|
|master|DejaVuSans.ttf|0.451248sec|
|this branch|DejaVuSans.ttf|0.077370sec|
|(reference)1.5.1|DejaVuSans.ttf|0.021442sec|
|master|GenShinGothic-Normal.ttf|1.706043sec|
|this branch|GenShinGothic-Normal.ttf|0.220942sec|
|(reference)1.5.1|GenShinGothic-Normal.ttf|0.064324sec|